#124 refactor: divide CalendarContainer inline styles to styled-components and change directory to src/Common/CalendarContainer
#124 refactor: change CustomCalendar and styles file directory in CalendarContainer

#124 test: change CustomCalendar import

# 124 fix: fix relative path in SchedulePage
