+++ src/components/Common/CalendarContainer/CalendarContainer.styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const CalendarContainerDiv = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: end;
+	margin-right: 2rem;
+	flex: 1;
+`;
