+++ src/pages/PersonalSchedulePage/PersonalSchedulePage.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import CalendarContainer from "@/components/Common/CalendarContainer";
 import scheduleType from "@/constants/calendar/scheduleType";
 
 import { LayoutMain } from "./PersonalSchedulePage.styles";
+import CalendarContainer from "../../components/Common/CalendarContainer/CalendarContainer";
 import Header from "../../components/Header/Header/Header";
 import PersonalTodoList from "../../components/PersonalTodoList/PersonalTodoList/PersonalTodoList";
